# Optimisation: Rush once finded