To my known its not possible to stop an NDRange once started, so this will try to rush ending by executing calculcation only if there isn't any result finded (because currently you must finish calculation of the whole aera to get the result).
The save is small if allrot is not (for an 50000 by 50000 by 60 aera, 4s was saved of 57s).
But with allroot time jump from 186s to 54s (same result as without allrot).

If we would to do that clean we are suposed to send small workload to the kernel and repeat if not finded (send 10s of work, if result finded print and quit else repeat), but that have a not negligible performance impact (it could be done by creating 2 workload so when one is finish the other is running but still a bit more slow).

Also I check on X and not an a new value because finding a result at X 0 is impossible (that full bedrock) and adding a new value cost a few µs per seconds.